Fix gait selection tests
