+++ march_gait_selection/CMakeLists.txt
@@ -24,6 +24,6 @@ install(DIRECTORY launch
 
 if (CATKIN_ENABLE_TESTING)
     find_package(rostest REQUIRED)
-    add_rostest(test/launch/testing_gait_selection.test)
     add_rostest(test/launch/testing_dynamic_gaits.test)
+    catkin_add_nosetests(test/run_tests.py)
 endif()
