+++ march_gait_selection/test/launch/testing_dynamic_gaits.test
@@ -1,10 +1,9 @@
 <launch>
     <param name="robot_description" textfile="$(find march_description)/urdf/march4.urdf"/>
-    <param name="balance" value="true" />
+    <param name="balance" value="true"/>
 
     <group ns="march">
-        <include file="$(find march_moveit)/launch/march_moveit.launch" />
-        <test test-name=" test_gait_selection" pkg="march_gait_selection" type="run_tests_dynamic_gaits"/>
+        <include file="$(find march_moveit)/launch/march_moveit.launch"/>
+        <test test-name="test_gait_selection" pkg="march_gait_selection" type="test_gait_balance.py"/>
     </group>
-
 </launch>
