+++ march_gait_selection/test/run_tests.py
@@ -6,7 +6,6 @@ from nosetests.test_gait_transition import TestTransitionTrajectory
 
 PKG = 'march_gait_selection'
 
-
 if __name__ == '__main__':
     rosunit.unitrun(PKG, 'test_gait_selection', TestGaitSelection)
     rosunit.unitrun(PKG, 'test_gait_transition', TestTransitionTrajectory)
