--- march_gait_selection/test/run_tests_dynamic_gaits
@@ -1,10 +1,0 @@
-#!/usr/bin/env python
-import rosunit
-
-from nosetests.test_gait_balance import TestBalanceGait
-
-PKG = 'march_gait_selection'
-
-
-if __name__ == '__main__':
-    rosunit.unitrun(PKG, 'test_gait_balance', TestBalanceGait)
